chore: rename ComboBox components
